+++ featuretools/primitives/standard/datetime_transform_primitives.py
@@ -14,8 +14,8 @@ from woodwork.logical_types import (
 
 from featuretools.primitives.base import TransformPrimitive
 from featuretools.primitives.utils import HolidayUtil
+from featuretools.types import Library
 from featuretools.utils import convert_time_units
-from featuretools.utils.gen_utils import Library
 
 
 class Age(TransformPrimitive):
