+++ featuretools/synthesis/deep_feature_synthesis.py
@@ -1,7 +1,7 @@
 import logging
 import warnings
 from collections import defaultdict
-from typing import Dict, List, Set, Tuple, Union, TypeVar
+from typing import Dict, List, Set, Tuple, TypeVar, Union
 
 from woodwork.column_schema import ColumnSchema
 from woodwork.logical_types import Boolean, BooleanNullable
@@ -28,7 +28,8 @@ from featuretools.primitives.options_utils import (
     generate_all_primitive_options,
     ignore_dataframe_for_primitive,
 )
-from featuretools.utils.gen_utils import Library, camel_and_title_to_snake
+from featuretools.types import Library, PrimitiveTypes
+from featuretools.utils.gen_utils import camel_and_title_to_snake
 
 logger = logging.getLogger("featuretools")
 
@@ -221,7 +222,7 @@ class DeepFeatureSynthesis(object):
                 if df_library in p.compatibility
             ]
         self.agg_primitives = sorted(
-            [check_primitive(p, "aggregation") for p in agg_primitives]
+            [check_primitive(p, PrimitiveTypes.AGGREGATION) for p in agg_primitives]
         )
 
         if trans_primitives is None:
@@ -231,21 +232,24 @@ class DeepFeatureSynthesis(object):
                 if df_library in p.compatibility
             ]
         self.trans_primitives = sorted(
-            [check_primitive(p, "transform") for p in trans_primitives]
+            [check_primitive(p, PrimitiveTypes.TRANSFORM) for p in trans_primitives]
         )
 
         self.where_primitives: List[PrimitiveBase]
         if where_primitives is None:
             where_primitives = [primitives.Count]
         self.where_primitives = sorted(
-            [check_primitive(p, "where") for p in where_primitives]
+            [check_primitive(p, PrimitiveTypes.WHERE) for p in where_primitives]
         )
 
         self.groupby_trans_primitives: List[PrimitiveBase]
         if groupby_trans_primitives is None:
             groupby_trans_primitives = []
         self.groupby_trans_primitives = sorted(
-            [check_primitive(p, "groupby transform") for p in groupby_trans_primitives]
+            [
+                check_primitive(p, PrimitiveTypes.GROUPBY_TRANSFORM)
+                for p in groupby_trans_primitives
+            ]
         )
 
         if primitive_options is None:
@@ -1098,10 +1102,13 @@ def instantiate_primitive(primitive: T) -> T:
     return primitive
 
 
-def check_primitive(primitive: Union[T, str], prim_type) -> T:
+def check_primitive(primitive: Union[T, str], prim_type: PrimitiveTypes) -> T:
 
     resolved_primitive: T
-    if prim_type == "transform" or prim_type == "groupby transform":
+    if (
+        prim_type == PrimitiveTypes.TRANSFORM
+        or prim_type == PrimitiveTypes.GROUPBY_TRANSFORM
+    ):
         prim_dict = primitives.get_transform_primitives()
         supertype = TransformPrimitive
         arg_name = (
@@ -1110,11 +1117,13 @@ def check_primitive(primitive: Union[T, str], prim_type) -> T:
             else "groupby_trans_primitives"
         )
         s = "a transform"
-    elif prim_type == "aggregation" or prim_type == "where":
+    elif prim_type == PrimitiveTypes.AGGREGATION or prim_type == PrimitiveTypes.WHERE:
         prim_dict = primitives.get_aggregation_primitives()
         supertype = AggregationPrimitive
         arg_name = (
-            "agg_primitives" if prim_type == "aggregation" else "where_primitives"
+            "agg_primitives"
+            if prim_type == PrimitiveTypes.AGGREGATION
+            else "where_primitives"
         )
         s = "an aggregation"
     else:
