+++ featuretools/tests/entityset_tests/test_last_time_index.py
@@ -7,7 +7,8 @@ from woodwork.logical_types import Categorical, Datetime, Integer
 
 from featuretools.entityset.entityset import LTI_COLUMN_NAME
 from featuretools.tests.testing_utils import to_pandas
-from featuretools.utils.gen_utils import Library, import_or_none
+from featuretools.types import Library
+from featuretools.utils.gen_utils import import_or_none
 
 ps = import_or_none("pyspark.pandas")
