+++ featuretools/tests/entityset_tests/test_ww_es.py
@@ -16,7 +16,8 @@ from woodwork.logical_types import (
 
 from featuretools.entityset.entityset import LTI_COLUMN_NAME, EntitySet
 from featuretools.tests.testing_utils import to_pandas
-from featuretools.utils.gen_utils import Library, import_or_none
+from featuretools.types import Library
+from featuretools.utils.gen_utils import import_or_none
 
 ps = import_or_none("pyspark.pandas")
