+++ featuretools/tests/synthesis/test_deep_feature_synthesis.py
@@ -51,7 +51,7 @@ from featuretools.primitives import (
 )
 from featuretools.synthesis import DeepFeatureSynthesis
 from featuretools.tests.testing_utils import feature_with_name, make_ecommerce_entityset
-from featuretools.utils.gen_utils import Library
+from featuretools.types import Library
 
 
 def test_makes_agg_features_from_str(es):
