+++ featuretools/types.py
@@ -1,3 +1,4 @@
+from enum import Enum
 from typing import Dict, Tuple
 
 import pandas as pd
@@ -10,3 +11,16 @@ T_SemanticTag = Dict[str, str]
 T_ES_Dataframe = Dict[
     str, Tuple[pd.DataFrame, str, str, T_LogicalType, T_SemanticTag, bool]
 ]
+
+
+class Library(Enum):
+    PANDAS = "pandas"
+    DASK = "Dask"
+    SPARK = "Spark"
+
+
+class PrimitiveTypes(Enum):
+    AGGREGATION = "aggregation"
+    TRANSFORM = "transform"
+    WHERE = "where"
+    GROUPBY_TRANSFORM = "groupby transform"
